requirements.txt
google-cloud-storage
google-api-core
requests
